#1 petita prova (#1)
* funciona

* Update

* add tests

* add inmemory

+++ 2.Shared/Models/Money.cs
@@ -2,7 +2,7 @@
 {
     public record Money
     {
-        public decimal Amount { get; init; } 
+        public double Amount { get; init; } 
         
         public Currencies Currency { get; init; }
